move performance related tests into own package
